refactoring Search page, memoizing, move SortSelect to separate directory, delete unused json files
