+++ src/components/App/App.jsx
@@ -13,7 +13,7 @@ import {
   PricesDrop,
   ProductCard,
 } from 'pages';
-import { Search } from 'pages/Search';
+import Search from 'pages/Search/Search';
 import { Route, Routes } from 'react-router-dom';
 import { Hidden } from './App.styled';
 import { CatalogLayout } from 'components/CatalogLayout/CatalogLayout';
@@ -30,10 +30,7 @@ export const App = () => {
             <Route index element={<Catalog />} />
             <Route path=":pet" element={<Pet />} />
             <Route path=":pet/:category" element={<Category />} />
-            <Route
-              path=":pet/:category/:productType"
-              element={<ProductType />}
-            />
+            <Route path=":pet/:category/:productType" element={<ProductType />} />
           </Route>
 
           {/*  <Route path="catalog/:pet" element={<Pet />}>
@@ -51,10 +48,7 @@ export const App = () => {
           <Route path="cart" element={<Cart />} />
           <Route path="success" element={<OrderSuccessful />} />
 
-          <Route
-            path="catalog/:pet/:category/:productType/:id"
-            element={<ProductCard />}
-          >
+          <Route path="catalog/:pet/:category/:productType/:id" element={<ProductCard />}>
             {/* <Route path="description" element={<ProductDescription />} />
             <Route path="composition" element={<ProductComposition />} />
             <Route path="comments" element={<ProductComments />} /> */}
@@ -65,10 +59,7 @@ export const App = () => {
 
           <Route path="aboutUs" element={<AboutUs />} />
           <Route path="contacts" element={<Contacts />} />
-          <Route
-            path="/pageUnderConstruction"
-            element={<PageUnderConstruction />}
-          />
+          <Route path="/pageUnderConstruction" element={<PageUnderConstruction />} />
         </Route>
         <Route path="*" element={<PageNotFound />} />
       </Routes>
