--- src/components/SearchBar/readme.md
@@ -1,7 +1,0 @@
-## Search component contains logic
-
-1. Validation (can send query with range from 3 to 255 characters) if message too short display
-   message (Занадто короткий запит) if message to long show message (Занадто довгий запит)
-2. Save state query
-3. query case-insensitive (toLowerCase)
-4. clear button clear query state
