+++ src/pages/Search/Search.jsx
@@ -1,34 +1,24 @@
-import { CardList, Pagination } from 'components';
+import React from 'react';
 import Loader from 'components/Loader/Loader';
 import { useEffect, useRef, useState } from 'react';
 import { useFetchSearchQuery } from 'redux/operations';
 import {
-  FoldedContainer,
   SearchAsideCatalog,
-  SearchBrands,
-  SearchCategoryList,
   SearchContainer,
-  SearchDescriptionSpan,
-  SearchDesriptionResults,
-  SearchFilter,
-  SearchQuery,
-  SearchWrapperCatalog,
   SortingContainer,
   TitleSearch,
   UpsideSearchContainer,
 } from './Search.styled';
-import { RightArrow } from 'components/Icons';
 import { useSelector } from 'react-redux';
-import { PriceSlider } from 'components/PriceSlider/PriceSlider';
-import React from 'react';
-import { Filter } from 'components/Filter/Filter';
-import { SortSelect } from 'components/Filter/SortSelect';
-import { theme } from 'styles';
+import { SortSelect } from 'components/SortSelect/SortSelect';
 import { selectSearchQueryStore, selectSortingTypeStore } from 'redux/selectors';
 import { NoSearch } from 'components/NoSearch/NoSearch';
 import { Notify } from 'notiflix';
+import SearchWrapper from './SearchWarapper';
+import SearchDescription from './SearchDescription';
+import SearchCategory from './SearchCategory';
 
-export const Search = () => {
+export default function Search() {
   const [productsList, setProductsList] = useState([]);
   const [currentPage, setCurrentPage] = useState(1);
   const [paginationData, setPaginationData] = useState({
@@ -42,7 +32,7 @@ export const Search = () => {
   });
   const [loadMoreProducts, setLoadMoreProducts] = useState([]); // Окремий стан для продуктів, завантажених через "Load More"
   const [loadMoreClicked, setLoadMoreClicked] = useState(false); // Окремий стан для слідкування за натисканням кнопки "Load More"
-  const [active, setActive] = useState({ price: false, brands: false });
+  // const [active, setActive] = useState({ price: false, brands: false });
   const abortControllerRef = useRef();
   const searchRef = useRef();
 
@@ -62,7 +52,6 @@ export const Search = () => {
     isLoading,
     isFetching,
     status,
-    currentData,
     isError,
   } = useFetchSearchQuery({
     query: searchQuery,
@@ -123,12 +112,6 @@ export const Search = () => {
     setCurrentPage(currentPage + 1);
   };
 
-  const handleClickToggle = e => {
-    active[e.currentTarget.attributes.name.value]
-      ? setActive({ ...active, [e.currentTarget.attributes.name.value]: false })
-      : setActive({ ...active, [e.currentTarget.attributes.name.value]: true });
-  };
-
   return (
     <div style={{ minHeight: '640px' }}>
       {error?.status >= 500 ? (
@@ -137,23 +120,11 @@ export const Search = () => {
         <Loader />
       ) : error?.status < 500 ? (
         <NoSearch status={status} />
-      ) : currentData ? (
+      ) : response ? (
         <>
           <UpsideSearchContainer>
             <TitleSearch>Результати пошуку</TitleSearch>
-            <SearchDesriptionResults>
-              <SearchDescriptionSpan>За запитом </SearchDescriptionSpan>
-              <SearchQuery>{searchRef.current.searchQuery}</SearchQuery>
-              <SearchDescriptionSpan> знайдено </SearchDescriptionSpan>
-              <SearchQuery>{searchRef.current.totalDocs} </SearchQuery>
-              <SearchDescriptionSpan>
-                {searchRef.current.totalDocs === 1
-                  ? 'товар'
-                  : searchRef.current.totalDocs < 5
-                  ? 'товари'
-                  : 'товарів'}
-              </SearchDescriptionSpan>
-            </SearchDesriptionResults>
+            <SearchDescription searchRef={searchRef} />
             <SortingContainer>
               <SortSelect />
             </SortingContainer>
@@ -161,76 +132,22 @@ export const Search = () => {
 
           <SearchContainer>
             <SearchAsideCatalog>
-              <SearchCategoryList>
-                <ul
-                  style={{
-                    display: 'flex',
-                    flexDirection: 'column',
-                    gap: '4px',
-                  }}
-                >
-                  <li key={1}>
-                    <SearchFilter active={active['price']}>
-                      <FoldedContainer
-                        active={active['price']}
-                        style={{
-                          backgroundColor: active['price']
-                            ? theme.colors.secGreen
-                            : theme.colors.beige,
-                        }}
-                        onClick={handleClickToggle}
-                        name="price"
-                      >
-                        <span>Ціна</span>
-                        <button name="price">
-                          <RightArrow direction={active['price'] ? 'rotate(90)' : 'rotate(-90)'} />
-                        </button>
-                      </FoldedContainer>
-                      <PriceSlider active={active['price']} />
-                    </SearchFilter>
-                  </li>
-                  <li key={2}>
-                    <SearchBrands activeBrands={active['brands']}>
-                      <FoldedContainer
-                        style={{
-                          backgroundColor: active['brands']
-                            ? theme.colors.secGreen
-                            : theme.colors.beige,
-                        }}
-                        onClick={handleClickToggle}
-                        name="brands"
-                      >
-                        <span>Бренди</span>
-                        <button name="brands">
-                          <RightArrow direction={active['brands'] ? 'rotate(90)' : 'rotate(-90)'} />
-                        </button>
-                      </FoldedContainer>
-                      <Filter active={active['brands']} />
-                    </SearchBrands>
-                  </li>
-                </ul>
-              </SearchCategoryList>
+              <SearchCategory />
             </SearchAsideCatalog>
-
-            <SearchWrapperCatalog>
-              <CardList
-                productsList={
-                  currentPage === 1
-                    ? productsList
-                    : loadMoreClicked
-                    ? loadMoreProducts
-                    : productsList
-                }
-              />
-              <Pagination
-                paginationData={paginationData}
-                onPageChange={onPageChange}
-                onAddPage={onAddPage}
-              />
-            </SearchWrapperCatalog>
+            <SearchWrapper
+              params={{
+                currentPage,
+                productsList,
+                loadMoreClicked,
+                loadMoreProducts,
+                paginationData,
+                onPageChange,
+                onAddPage,
+              }}
+            />
           </SearchContainer>
         </>
       ) : null}
     </div>
   );
-};
+}
